app/requirements.txt
# Suraksha App Dependencies
# All packages with minimum versions for reproducibility

# Core framework
streamlit>=1.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
mlflow>=2.8.0
joblib>=1.3.0

# RAG and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Translation
transformers>=4.35.0
torch>=2.0.0

# Additional utilities
python-dateutil>=2.8.2
